feat: new status psup & split scripts
:sparkles: "à publier (soumis à validation)"
