+++ server/package.json
@@ -32,10 +32,12 @@
     "psup:missing": "standalone=true node src/jobs/oneshot/psFormation/missingEtablissement.js",
     "psup:reconciliation": "standalone=true node src/jobs/oneshot/psFormation/reconciliationUpdate.js",
     "psup:2021": "standalone=true node src/jobs/psCoverage/pscoverage2021.js",
-    "psup:reference": "standalone=true node src/jobs/psReference/index.js",
+    "psup:reference": "standalone=true node src/jobs/reference/parcoursup/index.js",
+    "psup:pertinence": "standalone=true node src/jobs/pertinence/parcoursup/index.js",
     "affelnet:coverage": "standalone=true node src/jobs/afFormation/index.js",
     "affelnet:import": "standalone=true node src/jobs/oneshot/afFormation/index.js",
-    "affelnet:reference": "standalone=true node src/jobs/afReference/index.js",
+    "affelnet:reference": "standalone=true node src/jobs/reference/affelnet/index.js",
+    "affelnet:pertinence": "standalone=true node src/jobs/pertinence/affelnet/index.js",
     "affelnet:missing-cfd": "standalone=true node src/jobs/oneshot/afFormation/missingCfd.js",
     "affelnet:reconciliation": "standalone=true node src/jobs/afFormation/reconciliation.js",
     "opco:importer": "standalone=true node src/jobs/opcoImporter/index.js",
